requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
array-record==0.4.1
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
blinker==1.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
Django==4.1.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-tinymce==3.5.0
dm-tree==0.1.8
etils==1.5.2
executing==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gradio==4.19.2
gradio_client==0.10.1
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.4
importlib-resources==6.1.0
install==1.3.5
ipykernel==6.29.2
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.19.0
Jinja2==3.1.3
joblib==1.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.12.0
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.6.0
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.15
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.2.0
pooch==1.6.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
python-steam-api==1.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.5
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.36.3
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.6.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2022.7
urllib3==1.26.16
uvicorn==0.27.1
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0